# complete-api-docs.json
{
    "openapi":  "3.0.0",
    "paths":  {
...
                  "/api/study/session":  {
                                             "post":  {
                                                          "operationId":  "StudyController_start",
                                                          "summary":  "Start a new study session",
                                                          "description":  "Initiates a new spaced repetition study session with selected cards",
                                                          "parameters":  [

                                                                         ],
                                                          "requestBody":  {
                                                                              "required":  true,
                                                                              "description":  "Study session configuration including card selection criteria, session type, and optional time limits",
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/StartSessionDto"
                                                                                                                                  }
                                                                                                                   }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "201":  {
                                                                                        "description":  "Study session started successfully"
                                                                                    },
                                                                            "400":  {
                                                                                        "description":  "Invalid session parameters"
                                                                                    },
                                                                            "401":  {
                                                                                        "description":  "Unauthorized"
                                                                                    }
                                                                        },
                                                          "tags":  [
                                                                       "Study"
                                                                   ],
                                                          "security":  [
                                                                           {
                                                                               "bearer":  [

                                                                                          ]
                                                                           }
                                                                       ]
                                                      }
                                         },
                  "/api/study/session/{id}":  {
                                                  "get":  {
                                                              "operationId":  "StudyController_getSession",
                                                              "summary":  "Get session details (with answers)",
                                                              "description":  "Retrieves comprehensive details of a study session including progress and submitted answers",
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "required":  true,
                                                                                     "in":  "path",
                                                                                     "description":  "Study session ID",
                                                                                     "schema":  {
                                                                                                    "example":  "session-123-uuid-456",
                                                                                                    "type":  "string"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "Session details retrieved successfully"
                                                                                        },
                                                                                "401":  {
                                                                                            "description":  "Unauthorized"
                                                                                        },
                                                                                "404":  {
                                                                                            "description":  "Session not found"
                                                                                        }
                                                                            },
                                                              "tags":  [
                                                                           "Study"
                                                                       ],
                                                              "security":  [
                                                                               {
                                                                                   "bearer":  [

                                                                                              ]
                                                                               }
                                                                           ]
                                                          }
                                              },
                  "/api/study/session/{id}/end":  {
                                                      "post":  {
                                                                   "operationId":  "StudyController_end",
                                                                   "summary":  "End a session",
                                                                   "description":  "Completes a study session and calculates final statistics and performance metrics",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "description":  "Study session ID to end",
                                                                                          "schema":  {
                                                                                                         "example":  "session-123-uuid-456",
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Session ended successfully"
                                                                                             },
                                                                                     "400":  {
                                                                                                 "description":  "Session already ended"
                                                                                             },
                                                                                     "401":  {
                                                                                                 "description":  "Unauthorized"
                                                                                             },
                                                                                     "404":  {
                                                                                                 "description":  "Session not found"
                                                                                             }
                                                                                 },
                                                                   "tags":  [
                                                                                "Study"
                                                                            ],
                                                                   "security":  [
                                                                                    {
                                                                                        "bearer":  [

                                                                                                   ]
                                                                                    }
                                                                                ]
                                                               }
                                                  },
                  "/api/study/session/{id}/next":  {
                                                       "get":  {
                                                                   "operationId":  "StudyController_next",
                                                                   "summary":  "Get next card for this session",
                                                                   "description":  "Retrieves the next card to be studied using spaced repetition algorithm",
                                                                   "parameters":  [
                                                                                      {
                                                                                          "name":  "id",
                                                                                          "required":  true,
                                                                                          "in":  "path",
                                                                                          "description":  "Study session ID",
                                                                                          "schema":  {
                                                                                                         "example":  "session-123-uuid-456",
                                                                                                         "type":  "string"
                                                                                                     }
                                                                                      }
                                                                                  ],
                                                                   "responses":  {
                                                                                     "200":  {
                                                                                                 "description":  "Next card retrieved successfully"
                                                                                             },
                                                                                     "401":  {
                                                                                                 "description":  "Unauthorized"
                                                                                             },
                                                                                     "404":  {
                                                                                                 "description":  "Session not found or no more cards"
                                                                                             }
                                                                                 },
                                                                   "tags":  [
                                                                                "Study"
                                                                            ],
                                                                   "security":  [
                                                                                    {
                                                                                        "bearer":  [

                                                                                                   ]
                                                                                    }
                                                                                ]
                                                               }
                                                   },
                  "/api/study/session/{id}/answer":  {
                                                         "post":  {
                                                                      "operationId":  "StudyController_answer",
                                                                      "summary":  "Submit an answer for a card in the session",
                                                                      "description":  "Records user response and updates spaced repetition scheduling based on answer quality",
                                                                      "parameters":  [
                                                                                         {
                                                                                             "name":  "id",
                                                                                             "required":  true,
                                                                                             "in":  "path",
                                                                                             "description":  "Study session ID",
                                                                                             "schema":  {
                                                                                                            "example":  "session-123-uuid-456",
                                                                                                            "type":  "string"
                                                                                                        }
                                                                                         }
                                                                                     ],
                                                                      "requestBody":  {
                                                                                          "required":  true,
                                                                                          "description":  "User answer details including card ID, response content, correctness, and optional performance metrics",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "$ref":  "#/components/schemas/AnswerDto"
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      },
                                                                      "responses":  {
                                                                                        "201":  {
                                                                                                    "description":  "Answer submitted successfully"
                                                                                                },
                                                                                        "400":  {
                                                                                                    "description":  "Invalid answer data"
                                                                                                },
                                                                                        "401":  {
                                                                                                    "description":  "Unauthorized"
                                                                                                },
                                                                                        "404":  {
                                                                                                    "description":  "Session not found"
                                                                                                }
                                                                                    },
                                                                      "tags":  [
                                                                                   "Study"
                                                                               ],
                                                                      "security":  [
                                                                                       {
                                                                                           "bearer":  [

                                                                                                      ]
                                                                                       }
                                                                                   ]
                                                                  }
                                                     },
...
              },
    "info":  {
                 "title":  "Matrus API",
                 "description":  "AI-powered Leitner System API",
                 "version":  "1.0",
                 "contact":  {

                             }
             },
...
    "servers":  [

                ],
    "components":  {
                       "securitySchemes":  {
                                               "bearer":  {
                                                              "scheme":  "bearer",
                                                              "bearerFormat":  "JWT",
                                                              "type":  "http"
                                                          }
                                           },
                       "schemas":  {
...
                                       "StartSessionDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "description":  "Type of study session",
                                                                                               "enum":  [
                                                                                                            "review",
                                                                                                            "new",
                                                                                                            "mixed"
                                                                                                        ],
                                                                                               "example":  "mixed"
                                                                                           },
                                                                                  "limit":  {
                                                                                                "type":  "number",
                                                                                                "description":  "Maximum number of cards in this session",
                                                                                                "example":  20,
                                                                                                "minimum":  1,
                                                                                                "maximum":  500
                                                                                            }
                                                                              }
                                                           },
                                       "AnswerDto":  {
                                                         "type":  "object",
                                                         "properties":  {
                                                                            "cardId":  {
                                                                                           "type":  "string",
                                                                                           "description":  "ID of the card being answered",
                                                                                           "example":  "card-123-uuid-456"
                                                                                       },
                                                                            "correct":  {
                                                                                            "type":  "boolean",
                                                                                            "description":  "Whether the answer was correct",
                                                                                            "example":  true
                                                                                        },
                                                                            "userAnswer":  {
                                                                                               "type":  "string",
                                                                                               "description":  "User\u0027s actual answer text",
                                                                                               "example":  "A pleasant surprise"
                                                                                           },
                                                                            "timeSpent":  {
                                                                                              "type":  "number",
                                                                                              "description":  "Time spent on this card in seconds",
                                                                                              "example":  15,
                                                                                              "minimum":  0
                                                                                          },
                                                                            "selfRating":  {
                                                                                               "type":  "number",
                                                                                               "description":  "User self-assessment rating (1=very hard, 5=very easy)",
                                                                                               "example":  3,
                                                                                               "minimum":  1,
                                                                                               "maximum":  5
                                                                                           }
                                                                        },
                                                         "required":  [
                                                                          "cardId",
                                                                          "correct"
                                                                      ]
                                                     },
...
                                   }
                   }
}